move Web3bch.ts
