--- lib/index.d.ts
@@ -1,9 +1,0 @@
-import Providers from "./Providers";
-import Wallet from "./web3bch-wallet/Wallet";
-export default class Web3bch {
-    /**
-     * The current providers set.
-     */
-    providers: Providers;
-    wallet: Wallet;
-}
