+++ lib/web3bch/Web3bch.d.ts
@@ -1,5 +1,5 @@
 import Providers from "./Providers";
-import Wallet from "./web3bch-wallet/Wallet";
+import Wallet from "../web3bch-wallet/Wallet";
 export default class Web3bch {
     /**
      * The current providers set.
