+++ src/web3bch-wallet/Wallet.ts
@@ -4,7 +4,7 @@ import Utxo from "../web3bch-providers/entities/Utxo"
 import Network from "./entities/Network"
 import Destination from "./entities/Destination"
 import Output from "../web3bch-providers/entities/Output"
-import Providers from "../Providers"
+import Providers from "../web3bch/Providers"
 import IWalletProvider from "../web3bch-providers/IWalletProvider"
 import IllegalArgumentException from "./entities/IllegalArgumentException"
 import ProviderException from "./entities/ProviderException"
