+++ src/web3bch/Web3bch.ts
@@ -1,5 +1,5 @@
 import Providers from "./Providers"
-import Wallet from "./web3bch-wallet/Wallet"
+import Wallet from "../web3bch-wallet/Wallet"
 
 export default class Web3bch {
   /**
